requirements.txt
# Core audio processing
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0

# Real-time audio input
sounddevice>=0.4.6

# AI Agent Framework (preview)
agent-framework-azure-ai --pre

# OpenAI for GitHub models
openai>=1.30.0

# Speckit for acoustic analysis
# Note: Install from GitHub as per your requirement
# git+https://github.com/speckit-org/speckit.git

# Data processing and visualization
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0

# CLI interface
click>=8.1.0
rich>=13.0.0

# Optional: Jupyter for interactive analysis
jupyter>=1.0.0
ipywidgets>=8.0.0